Fix virtual runtime bug

+++ src/main/java/org/cap/TaskQueue.java
@@ -5,6 +5,6 @@ import java.util.PriorityQueue;
 public class TaskQueue extends PriorityQueue<Task> {
 
     public TaskQueue() {
-        super((task1, task2) -> Integer.compare(task1.virtualRuntime, task2.virtualRuntime));
+        super((task1, task2) -> Float.compare(task1.virtualRuntime, task2.virtualRuntime));
     }
 }
